[Pipeline Tool] Native Mac Version
