+++ Tools/Pipeline/Controls/PropertyGridTable.cs
@@ -242,10 +242,12 @@ namespace MonoGame.Tools.Pipeline
 
         private void PropertyGridTable_SizeChanged(object sender, EventArgs e)
         {
+#if WINDOWS
             SetWidth();
+#endif
 
 #if LINUX
-            // force size realocation
+            // force size reallocation
             drawable.Width = pixel1.Width - 2;
 
             foreach (var child in pixel1.Children)
@@ -256,7 +258,7 @@ namespace MonoGame.Tools.Pipeline
             drawable.Invalidate();
         }
 
-        private void SetWidth()
+        public void SetWidth()
         {
 #if WINDOWS
             var action = new Action(() =>
@@ -271,6 +273,9 @@ namespace MonoGame.Tools.Pipeline
 
             (drawable.ControlObject as System.Windows.Controls.Canvas).Dispatcher.BeginInvoke(action,
                 System.Windows.Threading.DispatcherPriority.ContextIdle, null);
+
+#elif MONOMAC
+            drawable.Width = Width; // TODO: Subtract sctollbar size
 #endif
         }
     }
