+++ Tools/Pipeline/Dialogs/AddItemDialog.eto.cs
@@ -16,8 +16,7 @@ namespace MonoGame.Tools.Pipeline
 
         private void InitializeComponent()
         {
-            if (!Global.Unix)
-                this.Height = 250;
+            Height = 250;
 
             layout1 = new DynamicLayout();
             layout1.DefaultSpacing = new Size(8, 8);
