+++ Tools/Pipeline/Global.cs
@@ -29,8 +29,10 @@ namespace MonoGame.Tools.Pipeline
         private static Dictionary<string, Image> _files;
         private static Image _fileMissing, _folder, _folderMissing;
 
+#if WINDOWS || LINUX
         private static Dictionary<string, Xwt.Drawing.Image> _xwtFiles;
         private static Xwt.Drawing.Image _xwtFileMissing, _xwtFolder, _xwtFolderMissing;
+#endif
 
         static Global()
         {
@@ -42,11 +44,13 @@ namespace MonoGame.Tools.Pipeline
             _folder = Bitmap.FromResource("TreeView.Folder.png");
             _folderMissing = Bitmap.FromResource("TreeView.FolderMissing.png");
 
+#if WINDOWS || LINUX
             _xwtFiles = new Dictionary<string, Xwt.Drawing.Image>();
             _xwtFiles.Add(".", Xwt.Drawing.Image.FromResource("TreeView.File.png"));
             _xwtFileMissing = Xwt.Drawing.Image.FromResource("TreeView.FileMissing.png");
             _xwtFolder = Xwt.Drawing.Image.FromResource("TreeView.Folder.png");
             _xwtFolderMissing = Xwt.Drawing.Image.FromResource("TreeView.FolderMissing.png");
+#endif
 
             PlatformInit();
         }
@@ -91,6 +95,7 @@ namespace MonoGame.Tools.Pipeline
             return icon;
         }
 
+#if WINDOWS || LINUX
         public static Xwt.Drawing.Image GetXwtDirectoryIcon(bool exists)
         {
             return exists ? _xwtFolder : _xwtFolderMissing;
@@ -119,6 +124,7 @@ namespace MonoGame.Tools.Pipeline
             _xwtFiles.Add(ext, icon);
             return icon;
         }
+#endif
 
         public static Image GetEtoIcon(string resource)
         {
@@ -132,6 +138,7 @@ namespace MonoGame.Tools.Pipeline
             return Icon.FromResource(resource);
         }
 
+#if WINDOWS || LINUX
         public static Xwt.Drawing.Image GetXwtIcon(string resource)
         {
 #if LINUX
@@ -143,6 +150,7 @@ namespace MonoGame.Tools.Pipeline
             
             return Xwt.Drawing.Image.FromResource(resource);
         }
+#endif
     }
 }
