+++ Tools/Pipeline/MainWindow.eto.cs
@@ -70,8 +70,6 @@ namespace MonoGame.Tools.Pipeline
 
             Content = splitterHorizontal;
 
-            projectControl.TreeView.RowActivated += CmdOpenItem_Executed;
-
             cmdNew.Executed += CmdNew_Executed;
             cmdOpen.Executed += CmdOpen_Executed;
             cmdClose.Executed += CmdClose_Executed;
@@ -141,7 +139,7 @@ namespace MonoGame.Tools.Pipeline
             cmdSaveAs.Image = Global.GetEtoIcon("Commands.SaveAs.png");
 
             cmdExit = new Command();
-            cmdExit.MenuText = "Exit";
+            cmdExit.MenuText = (Global.Unix && !Global.Linux) ? "Quit" : "Exit";
             cmdExit.Shortcut = Application.Instance.CommonModifier | Keys.Q;
 
             // Edit Commands
@@ -254,6 +252,10 @@ namespace MonoGame.Tools.Pipeline
         private void InitalizeMenu()
         {
             menubar = new MenuBar();
+#if MONOMAC
+            Menu = menubar;
+            Menu.Items.Clear(); // HACK: Eto.Forms adds some pointless items on Mac
+#endif
 
             menuFile = new ButtonMenuItem();
             menuFile.Text = "File";
@@ -353,15 +355,15 @@ namespace MonoGame.Tools.Pipeline
             ToolBar.Items.Add(cmdNew);
             ToolBar.Items.Add(cmdOpen);
             ToolBar.Items.Add(cmdSave);
-            ToolBar.Items.Add(new SeparatorToolItem());
+            ToolBar.Items.Add(new SeparatorToolItem { Type = SeparatorToolItemType.Divider });
             ToolBar.Items.Add(cmdUndo);
             ToolBar.Items.Add(cmdRedo);
-            ToolBar.Items.Add(new SeparatorToolItem());
+            ToolBar.Items.Add(new SeparatorToolItem { Type = SeparatorToolItemType.Divider });
             ToolBar.Items.Add(cmdNewItem);
             ToolBar.Items.Add(cmdExistingItem);
             ToolBar.Items.Add(cmdNewFolder);
             ToolBar.Items.Add(cmdExistingFolder);
-            ToolBar.Items.Add(new SeparatorToolItem());
+            ToolBar.Items.Add(new SeparatorToolItem { Type = SeparatorToolItemType.Divider });
             ToolBar.Items.Add(toolBuild);
             ToolBar.Items.Add(toolRebuild);
             ToolBar.Items.Add(toolClean);
